@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imatic\Bundle\ConfigBundle\DependencyInjection\Compiler;
 
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
@@ -7,18 +7,14 @@ use Symfony\Component\DependencyInjection\Reference;
 
 class CompilerPass implements CompilerPassInterface
 {
-    /**
-     * @param ContainerBuilder $container
-     */
+    public const TAG = 'imatic_config.provider';
+
     public function process(ContainerBuilder $container)
     {
         $definition = $container->getDefinition('imatic_config.config_manager');
 
-        foreach ($container->findTaggedServiceIds('imatic_config.provider') as $id => $parameters) {
-            $definition->addMethodCall('registerProvider', [
-                new Reference($id),
-                isset($parameters[0]['alias']) ? $parameters[0]['alias'] : $id
-            ]);
+        foreach ($container->findTaggedServiceIds(self::TAG) as $id => $parameters) {
+            $definition->addMethodCall('registerProvider', [new Reference($id), $parameters[0]['alias'] ?? 'config']);
         }
     }
 }
