@@ -1,70 +1,52 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imatic\Bundle\ConfigBundle\Command;
 
-use Imatic\Bundle\ConfigBundle\Config\ConfigManager;
-use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
-use Symfony\Component\Console\Helper\TableHelper;
+use Imatic\Bundle\ConfigBundle\Config\ConfigManagerInterface;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Translation\Translator;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
-class ConfigListCommand extends ContainerAwareCommand
+class ConfigListCommand extends Command
 {
-    /**
-     * {@inheritDoc}
-     */
+    private ConfigManagerInterface $configManager;
+
+    private TranslatorInterface $translator;
+
+    public function __construct(ConfigManagerInterface $configManager, TranslatorInterface $translator)
+    {
+        $this->configManager = $configManager;
+        $this->translator = $translator;
+
+        parent::__construct();
+    }
+
     protected function configure()
     {
         $this
             ->setName('imatic:config:list')
             ->setDescription('Lists all available config variables.')
-            ->addArgument('filter', InputArgument::OPTIONAL, 'Filters valiable keys by the given filter.');
+            ->addArgument('filter', InputArgument::OPTIONAL, 'Filters variable keys by the given filter.');
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $configManager = $this->getConfigManager();
-        $table = $this->getTableHelper();
+        $table = new Table($output);
         $table->setHeaders(['key', 'value', 'type', 'description']);
-        $translator = $this->getTranslator();
 
-        foreach ($configManager->getViewValues($input->getArgument('filter')) as $key => $viewValue) {
+        foreach ($this->configManager->getViewValues($input->getArgument('filter')) as $key => $viewValue) {
             $table->addRow([
                 $key,
                 $viewValue,
-                $configManager->getDefinition($key)->getType(),
-                $translator->trans($key, [], 'config')
+                $this->configManager->getDefinition($key)->getType(),
+                $this->translator->trans($key, [], 'config'),
             ]);
         }
 
-        $table->render($output);
-    }
+        $table->render();
 
-    /**
-     * @return ConfigManager
-     */
-    private function getConfigManager()
-    {
-        return $this->getContainer()->get('imatic_config.config_manager');
-    }
-
-    /**
-     * @return TableHelper
-     */
-    private function getTableHelper()
-    {
-        return $this->getHelper('table');
-    }
-
-    /**
-     * @return Translator
-     */
-    private function getTranslator()
-    {
-        return $this->getContainer()->get('translator');
+        return 0;
     }
 }
