@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imatic\Bundle\ConfigBundle\Exception;
 
 class InvalidValueException extends \RuntimeException implements ExceptionInterface
