@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imatic\Bundle\ConfigBundle;
 
 use Imatic\Bundle\ConfigBundle\DependencyInjection\Compiler\CompilerPass;
@@ -7,9 +7,6 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class ImaticConfigBundle extends Bundle
 {
-    /**
-     * @param ContainerBuilder $container
-     */
     public function build(ContainerBuilder $container)
     {
         $container->addCompilerPass(new CompilerPass());
