@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imatic\Bundle\ConfigBundle\Provider;
 
 interface ProviderInterface
@@ -6,5 +6,5 @@ interface ProviderInterface
     /**
      * @return Definition[]
      */
-    public function getDefinitions();
+    public function getDefinitions(): array;
 }
